Fix integer overflow issues in problems F and G
Changed sum in f.cpp and map/ans types in g.cpp from int to long long to prevent potential overflow when handling large values.

+++ icpc/2025nanjing/g.cpp
@@ -20,8 +20,8 @@ void solve() {
     sort(b.begin() + 1, b.end());
     int q;
     cin >> q;
-    map<int, vector<int>> mp;
-    vector<int> ans(q + 1);
+    map<long long, vector<int>> mp;
+    vector<long long> ans(q + 1);
     for (int i = 1; i <= q; i++) {
         cin >> ans[i];
         mp[ans[i]].push_back(i);
